refactor: corrected namespace for various sandbox dtos
they were moved into separate folder earlyer, but namespace was not updated
