+++ src/Sepes.Functions/ResourceOperationsQueueFunction.cs
@@ -2,7 +2,7 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 using Microsoft.WindowsAzure.Storage.Queue;
 using Newtonsoft.Json;
-using Sepes.Infrastructure.Dto;
+using Sepes.Infrastructure.Dto.Sandbox;
 using Sepes.Infrastructure.Model.Config;
 using Sepes.Infrastructure.Service.Interface;
 using System.Threading.Tasks;
