+++ src/Sepes.Infrastructure/Dto/Sandbox/SandboxDatasetsDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Sepes.Infrastructure.Dto.Sandbox
+{
+    public class SandboxDatasetsDto
+    {
+    }
+}
