+++ src/Sepes.Infrastructure/Dto/Sandbox/SandboxResourceDto.cs
@@ -1,4 +1,4 @@
-﻿
+﻿using Sepes.Infrastructure.Dto.Sandbox;
 using System;
 using System.Collections.Generic;
