+++ src/Sepes.Infrastructure/Dto/Sandbox/SandboxResourceLightDto.cs
@@ -1,6 +1,6 @@
 ﻿using Sepes.Infrastructure.Dto.Interfaces;
 
-namespace Sepes.Infrastructure.Dto
+namespace Sepes.Infrastructure.Dto.Sandbox
 {
     public class SandboxResourceLightDto : IHasLinkToExtSystem
     {
