+++ src/Sepes.Infrastructure/Dto/StudyDto.cs
@@ -1,4 +1,5 @@
-﻿using Sepes.Infrastructure.Interface;
+﻿using Sepes.Infrastructure.Dto.Sandbox;
+using Sepes.Infrastructure.Interface;
 using System.Collections.Generic;
 
 namespace Sepes.Infrastructure.Dto
