+++ src/Sepes.Infrastructure/Model/Automapper/AutoMappingConfigs.cs
@@ -4,6 +4,7 @@ using Microsoft.Azure.Management.ResourceManager.Fluent.Core;
 using Sepes.Infrastructure.Constants;
 using Sepes.Infrastructure.Dto;
 using Sepes.Infrastructure.Dto.Azure;
+using Sepes.Infrastructure.Dto.Sandbox;
 using Sepes.Infrastructure.Dto.VirtualMachine;
 using Sepes.Infrastructure.Util;
 using System.Linq;
