+++ src/Sepes.Infrastructure/Service/Interface/IProvisioningQueueService.cs
@@ -1,4 +1,4 @@
-﻿using Sepes.Infrastructure.Dto;
+﻿using Sepes.Infrastructure.Dto.Sandbox;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
