+++ src/Sepes.Infrastructure/Service/Interface/ISandboxResourceOperationService.cs
@@ -1,4 +1,4 @@
-﻿using Sepes.Infrastructure.Dto;
+﻿using Sepes.Infrastructure.Dto.Sandbox;
 using Sepes.Infrastructure.Model;
 using System.Collections.Generic;
 using System.Threading.Tasks;
