+++ src/Sepes.Infrastructure/Util/AzureResourceTagsFactory.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Sepes.Infrastructure.Dto;
+using Sepes.Infrastructure.Dto.Sandbox;
 using Sepes.Infrastructure.Model.Config;
 using System;
 using System.Collections.Generic;
